fix: changes due to the code review
